refactor: GameEventListener 수정
- 구현체 제거
- Activity에서 상속받도록 수정
